Fix automated testing errors
ps-id: 594dbbab-2930-487b-ba98-d200f2f547e2
