+++ pylintrc
@@ -82,7 +82,7 @@ persistent=yes
 
 # Minimum Python version to use for version dependent checks. Will default to
 # the version used to run pylint.
-py-version=3.9
+py-version=3.8
 
 # Discover python modules and packages in the file system subtree.
 recursive=no
